HBX-1677: Move interface 'org.hibernate.tool.hbm2x.MetaAttributeConstants' to 'org.hibernate.tool.internal.export.pojo.MetaAttributeConstants'
